+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 1.1.0
+**This version includes Breaking Changes**
+- Changed type of StarRailUser#forgottenHall to ForgottenHallInfo.
 # 1.0.2
 - Updated image cdns.
 - Added Eidolon#picture, CombatType#icon, and Path#icon.